requirements.txt
fastapi
pydantic
uvicorn
streamlit
scikit-learn
matplotlib
pandas
python-dotenv
